+++ src/inspect_ai/model/_providers/grok.py
@@ -428,14 +428,14 @@ class GrokAPI(ModelAPI):
             gconfig["response_format"] = "json_object"
 
         # note that grok-3-mini is the only model which supports a reasoning effort parameter
-        if config.reasoning_effort is not None and self.is_grok_3_mini():
+        if config.reasoning_effort is not None and (
+            self.is_grok_3_mini() or self.is_grok_4()
+        ):
             match config.reasoning_effort:
-                case "none":
-                    raise ValueError(
-                        "Grok models do not support 'none' for reasoning effort."
-                    )
                 case "minimal" | "low":
                     gconfig["reasoning_effort"] = "low"
+                case "medium":
+                    gconfig["reasoning_effort"] = "medium"
                 case "medium" | "high" | "xhigh" | "max":
                     gconfig["reasoning_effort"] = "high"